+++ internal/configuration/config.template.yml
@@ -17,7 +17,7 @@ port: 9091
 ## They should be in base64 format, and have one of the following extensions: *.cer, *.crt, *.pem.
 # certificates_directory: /config/certificates
 
-## The theme to display: light, dark, grey, auto.
+## The theme to display: light, dark, grey.
 theme: light
 
 ##
@@ -588,9 +588,12 @@ notifier:
 
     ## The lifespans configure the expiration for these token types.
     # access_token_lifespan: 1h
-    # authorize_code_lifespan: 1h
+    # authorize_code_lifespan: 1m
     # id_token_lifespan: 1h
     # refresh_token_lifespan: 720h
+
+    ## Enables additional debug messages.
+    # enable_client_debug_messages: false
 
     ## Clients is a list of known clients and their configuration.
     # clients:
