+++ internal/handlers/handler_oidc_jwks.go
@@ -9,7 +9,7 @@ import (
 func oidcJWKs(ctx *middlewares.AutheliaCtx) {
 	ctx.SetContentType("application/json")
 
-	if err := json.NewEncoder(ctx).Encode(ctx.Providers.OpenIDConnect.Store.KeyManager.GetKeySet()); err != nil {
+	if err := json.NewEncoder(ctx).Encode(ctx.Providers.OpenIDConnect.KeyManager.GetKeySet()); err != nil {
 		ctx.Error(err, "failed to serve jwk set")
 	}
 }
