+++ internal/oidc/client.go
@@ -33,23 +33,6 @@ func NewClient(config schema.OpenIDConnectClientConfiguration) (client *Internal
 	return client
 }
 
-// InternalClient represents the client internally.
-type InternalClient struct {
-	ID            string              `json:"id"`
-	Description   string              `json:"-"`
-	Secret        []byte              `json:"client_secret,omitempty"`
-	RedirectURIs  []string            `json:"redirect_uris"`
-	GrantTypes    []string            `json:"grant_types"`
-	ResponseTypes []string            `json:"response_types"`
-	Scopes        []string            `json:"scopes"`
-	Audience      []string            `json:"audience"`
-	Public        bool                `json:"public"`
-	Policy        authorization.Level `json:"-"`
-
-	// These are the OpenIDConnect Client props.
-	ResponseModes []fosite.ResponseModeType `json:"response_modes"`
-}
-
 // IsAuthenticationLevelSufficient returns if the provided authentication.Level is sufficient for the client of the AutheliaClient.
 func (c InternalClient) IsAuthenticationLevelSufficient(level authentication.Level) bool {
 	return authorization.IsAuthLevelSufficient(level, c.Policy)
