+++ internal/oidc/hasher.go
@@ -5,12 +5,8 @@ import (
 	"crypto/subtle"
 )
 
-// AutheliaHasher implements the fosite.Hasher interface without an actual hashing algo.
-type AutheliaHasher struct {
-}
-
 // Compare compares the hash with the data and returns an error if they don't match.
-func (h AutheliaHasher) Compare(ctx context.Context, hash, data []byte) (err error) {
+func (h AutheliaHasher) Compare(_ context.Context, hash, data []byte) (err error) {
 	if subtle.ConstantTimeCompare(hash, data) == 0 {
 		return errPasswordsDoNotMatch
 	}
@@ -19,6 +15,6 @@ func (h AutheliaHasher) Compare(ctx context.Context, hash, data []byte) (err err
 }
 
 // Hash creates a new hash from data.
-func (h AutheliaHasher) Hash(ctx context.Context, data []byte) (hash []byte, err error) {
+func (h AutheliaHasher) Hash(_ context.Context, data []byte) (hash []byte, err error) {
 	return data, nil
 }
